Fixed a crash when trying to load a model with more than 32767 vertices.

+++ Source/Core/GZBuilder/md3/GZModel.cs
@@ -12,13 +12,13 @@ namespace CodeImp.DoomBuilder.GZBuilder.MD3
         internal byte NUM_MESHES; //can't be greater than 255, can it?
         internal Vector3[] BoundingBox;
         internal List<IndexBuffer> Indeces2D;
-        internal List<short> NumIndeces2D;
+        internal List<int> NumIndeces2D;
 
         internal GZModel() {
             Meshes = new List<Mesh>();
             Textures = new List<Texture>();
             Indeces2D = new List<IndexBuffer>();
-            NumIndeces2D = new List<short>();
+            NumIndeces2D = new List<int>();
         }
     }
